requirements.txt
# Raspberry Pi 5 AI System Dependencies - Pydro

# MQTT Client
paho-mqtt==1.6.1

# Image Processing & ML
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
tflite-runtime==2.14.0

# Twilio SMS
twilio==8.10.0

# Email (weekly reports)
secure-smtplib==0.1.1

# API Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Data Analysis
pandas==2.1.3

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Logging
colorlog==6.8.0

# Alexa Integration
Flask==3.0.0
Flask-Ask==0.9.8
cryptography==41.0.7

# System monitoring
psutil==5.9.6


# Optional: For local testing without RPi hardware
# picamera2  # Install via apt on Raspberry Pi OS only
